moved UserNotFoundException to exceptions folder
